% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sing_day.R
\name{GameScore}
\alias{GameScore}
\title{Produces game score values}
\usage{
GameScore(df)
}
\arguments{
\item{df}{A data frame containing information about NBA}
\value{
A vector of game score values
}
\description{
Produces the game score of the data entered
}
