% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize_gift.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Takes a noun and makes it plural}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{gift}{A string or vector of strings}
}
\value{
A string or vector of strings with the pluralized words
}
\description{
Takes a noun and makes it plural
}
