% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize_gift.R
\name{playerid}
\alias{playerid}
\title{removes id char from player}
\usage{
many_sample_means(dataset)
}
\arguments{
\item{dataset}{A datset}
}
\value{
a vector.
}
\description{
removes id char from player
}
