% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phrase.R
\name{sample_mean}
\alias{sample_mean}
\title{not blank}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{num}{An integer}

\item{num_word}{A string corresponding to the integer}

\item{item}{A string}

\item{verb}{A string}

\item{adjective}{A string}

\item{location}{A string}
}
\value{
A string containing the words in grammatical order.
}
\description{
Puts the various parts of speech together into a full phrase.
}
