% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sing_day.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Produces the string for one day of the song.}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{dataset}{A data frame containing information about gifts}

\item{line}{The number of the line for the day you want to sing about}

\item{phrase_col}{The variable name for the column in the dataset that
contains the gift phrases}
}
\value{
A string singing the line of the song with all gifts for the given day.
}
\description{
Produces the string for one day of the song.
}
